add secure annotation to wes service and permit run in controller
