+++ src/main/java/org/icgc_argo/workflow/search/graphql/RunDataFetchers.java
@@ -26,7 +26,7 @@ import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.icgc_argo.workflow.search.model.graphql.Run;
 import org.icgc_argo.workflow.search.model.graphql.Task;
-import org.icgc_argo.workflow.search.service.RunService;
+import org.icgc_argo.workflow.search.service.graphql.RunService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
